==> scripts/final/00_archetype-analysis_forest_gain-data-prep.R <==


#---Note---------
# This data is downloaded from 
# https://storage.googleapis.com/earthenginepartners-hansen/GFC-2023-v1.11/download.html
# one has to download the desire 10 x 10 degree boxes
# For the contiguous US this includes 
# 50N, 130W; 50N, 120W; 50N, 110W; 50N, 100W; 50N, 90W; 50N, 80W; 50N, 70W
# 40N, 130W; 40N, 120W; 40N, 110W; 40N, 100W; 40N, 90W; 40N, 80W
# 30N, 120W; 30N, 110W; 30N, 100W; 30N, 90W; 30N, 80W

# Download the data
